Move utility file for XPU and CUDA visibility
