+++ aten/src/ATen/native/cuda/ScaledBlas.cpp
@@ -11,7 +11,7 @@
 #include <ATen/ExpandUtils.h>
 #include <ATen/OpMathType.h>
 #include <ATen/TensorUtils.h>
-#include <ATen/native/ScaledBlasUtils.h>
+#include <ATen/ScaledBlasUtils.h>
 #include <ATen/cuda/tunable/Tunable.h>
 #include <ATen/cuda/tunable/TunableGemm.h>
 #include <ATen/native/Resize.h>
