+++ aten/src/ATen/native/mkldnn/xpu/ScaledBlas.cpp
@@ -3,7 +3,7 @@
 #include <ATen/WrapDimUtilsMulti.h>
 #include <ATen/ceil_div.h>
 #include <ATen/native/Resize.h>
-#include <ATen/native/ScaledBlasUtils.h>
+#include <ATen/ScaledBlasUtils.h>
 #include <ATen/native/mkldnn/xpu/detail/oneDNN.h>
 #include <ATen/native/xpu/Blas.h>
 #include <torch/library.h>
